requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
transformers==4.35.0
torch==2.1.0
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
python-dotenv==1.0.0
werkzeug==3.0.1

